% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{area_quantile}
\alias{area_quantile}
\title{Area bajo la curva de los graficos de densidad dividido por percentiles.}
\usage{
area_quantile(
  x,
  len = 4,
  color_area = NA,
  col_border = "black",
  titulo = "",
  dureza = 2,
  ...
)
}
\arguments{
\item{x}{Vector numérico del cual se desea obtener su gráfico de densidad.}

\item{len}{Parámetro numérico que equivale al numero de cortes que se desea hacer, es así que si
se desea calcular los elementos divididos por terciles, `len` toma el valor de 3. len tomo el
valor de 4 por defecto.}

\item{color_area}{Elemento numérico o vector caracter que especifica el color de las sub áreas.
Su valor por defecto es NA.}

\item{col_border}{Elemento caracter para especificar el color de los bordes que dividen las
sub áreas. El color por defecto es "black".}

\item{titulo}{Elemento caracter que se colocará en el título del gráfico.
Su valor por defecto es "".}

\item{dureza}{Elemento numérico que indica el grosor de la línea que divide las sub áreas.
Es el mismo argumento que `lwd` de la función plot. Su valor por defecto es 2.}

\item{...}{Otros argumentos que se pueden usar en la función plot. Estos argumentos afectarán
al plot, pero no a las sub áreas. Por ejemplo, puede usar `xlab`, `font`, `sub`, `bty`, etc. Pero no
puede usar el argumento `main`, ya que ya está especificado con el argumento `título`.}
}
\description{
Devuelve sub áreas de los gráficos de densidad divididos por percentiles.
Por ejemplo, terciles, cuartiles, quintiles, deciles, centiles, etc.
}
\details{
Para el cálculo de los percentiles se usa como base el vector definido en el argumento `x`
con esos puntos de corte se procede a dividir los valores del eje de las abscisas de la densidad
del vector `x` (la densidad se calcula con la función `density` de R base). Cabe precisar que al
ser transformado el vector `x` por la función de densidad los valores
de `x` se ajustan, por lo que no serán los mismos. Por tanto, si se desea determinar el número
de elementos que se encuentran por cada percentil se debería de calcular en el mismo vector `x`,
antes de transformar estos valores por la función de densidad.
Lo recomendable es que `color_area` sea un vector caracter con el mismo número de elementos
que `len`.
}
\examples{
# Vamos a usar la tabla de datos iris.
# Graficando las sub areas de la funcion de densidad de la variable Sepal.Length por quintiles
# y usando el argumento bty para quitar el marco del grafico.

area_quantile(iris$Sepal.Length, len = 5, bty="n")
}
