% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{dpto_peru}
\alias{dpto_peru}
\title{Ubigeo Peru}
\usage{
dpto_peru(x)
}
\arguments{
\item{x}{un elemento o vector character.}
}
\description{
Convertir el código numérico de ubigeo (código de ubicación geográfica del Perú)
a nivel departamental.
}
\details{
El o los elementos de `x` pueden ser los 6 dígitos que brinda el INEI.
Incluso el vector `x` puede tener entre 1 y 2 caracteres, por tanto no es necesario expresar
el ubigeo con un `0` antes de los primeros 9 departamentos.
}
\examples{
dpto_peru(c("1", "01", "010405", "12", "120902", "9", "090302"))
# "AMAZONAS" "AMAZONAS"  "AMAZONAS" "JUNIN" "JUNIN" "HUANCAVELICA" "HUANCAVELICA"
}
