% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{estadisticos}
\alias{estadisticos}
\title{Estadisticos}
\usage{
estadisticos(x, miss = T)
}
\arguments{
\item{x}{Vector numérico.}

\item{miss}{valor lógico que por defecto toma el valor de TRUE, que indica que no se debe
de considerar a los NAs en el cálculo.}
}
\description{
Devuelve lo principales estadísticos como: la media, la mediana, la desviación estándar
el mínimo, máximo, coeficiente de variación y el número de observaciones.
}
\details{
el vector `x` tiene que ser numérico, sino no se podrá calcular los estadísticos.
El coeficiente de variación es mayor a 20% indica que deberíamos tomar a la mediana como estadístico
de tendencia central, y si es menor el mejor estadísitico de tendencia central será la media.
Los resultados están configurados para que el número de decimales sea de 3. Asimismo, si usa
algunas de las funciones de la familia `apply` el nombre de los estadísticos no se mostrará.
En todo caso si quiere hacer el cálculo para varios vectores a la vez se recomienda usar la siguiente
sintaxis: `do.call(cbind,lapply(df, estadisticos))`, donde `df` es el data frame que contiene todos
los vectores del cual se desea calcular los estadísticos.
}
\examples{
set.seed(2020)
estadisticos(sample(1:100, 400,T))
#                   [,1]
# MEDIA           49.862
# MEDIANA         48.000
# DESV.STAND.     28.487
# MINIMO           1.000
# MAXIMO         100.000
# COEF.VARIACION   0.571
# N              400.000
}
