% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{grubbs_total}
\alias{grubbs_total}
\title{Test de grubbs que sirve para determinar si uno o varios elementos del
vector son outliers.}
\usage{
grubbs_total(x, alpha = 0.05, num_colas = 2)
}
\arguments{
\item{x}{Vector numérico de donde se determina si uno o varios elemento son outliers o no.}

\item{alpha}{Es el nivel de significancia con el que se desea determinar el valor crítico de
Grubbs. El valor por defecto es de 0.05.}

\item{num_colas}{Valor numérico que toma el valor de 1 si se desea calcular el valor
crítico a una sola cola. Si toma el valor de 2 se calcula el valor crítico a dos colas.
El valor por defecto es de 2 colas.}
}
\description{
Devuelve un vector de elementos lógicos, en donde si es TRUE un elemento
significa que ese elemento es outliers según el método de Grubss.
}
\details{
Para calcular el estadístico de Grubbs se usa:
\deqn{G=\frac{\left|x_i-\bar{x}\right|}{\sigma}}
Asimismo, la elección que se hace para saber si el elemento es outliers o no se hace con la
siguiente comparación.
\deqn{G_t>G_{tc} \rightarrow TRUE \rightarrow \textrm{Es outliers}}
\deqn{G_t>G_{tc} \rightarrow FALSE \rightarrow \textrm{No es outliers}}
}
\examples{
# Definimos primero el siguiente vector.
set.seed(2020)
x<-c(25, sample(1:10, 30, T), 21)
Aplicando el test para poder ver que elementos son outliers.
grubbs_total(x)
#  [1]  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#  [12] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#  [22] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
}
