% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{limpiecito}
\alias{limpiecito}
\title{Limpieza de caracteres o strings.}
\usage{
limpiecito(x, tipo = "no_et_mayu")
}
\arguments{
\item{x}{Un elemento o vector character.}

\item{tipo}{Tiene 4 opciones: no_et_minu (se quitan espacios, tiles y todo a minúsculas),
 no_et_mayu (se quitan espacios, tiles y todo a mayúsculas),
 no_et (se quitan espacios, tiles y se deja en el formato que estaba el texto) y no_e (solo se quitan espacios)
se mostrarán en mayúsculas y si toma el valor de FALSE los resultados se mostrarán en
minúsculas.}
}
\description{
Es una función para limpiar espacios por demás al inicio, final o intermedio. Asimismo
elimina tildes y arroja los resultados en letras mayúsculas o minúsculas.
}
\details{
Esta función se usará para la limpieza de base de datos.
}
\examples{
limpiecito(c("    raul", " yo uso    R"))
# "RAUL" "YO USO R"
}
