% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{n_percentil}
\alias{n_percentil}
\title{Numero de elementos por percentil.}
\usage{
n_percentil(x, len = 4, p_corte = T)
}
\arguments{
\item{x}{Vector numérico.}

\item{len}{parámetro numérico que equivale al numero de cortes que se desea hacer, es así que si
se desea calcular los elementos divididos por terciles, `len` toma el valor de 3. len tomo el
valor de 4 por defecto.}

\item{p_corte}{valor lógico que toma el valor de TRUE por defecto, que indica que en los resultados
se muestre los puntos de cortes de los percentiles. Si es FALSE, en el resultado se mostrará por
partes, es así que si se escoge cuartiles, se mostrará el nombre de las filas del uno al 4, que hace
referencia a las partes en las que se dividió.}
}
\description{
Devuelve el número de elementos que se encuentran por cada percentil especificado
Es así que se puede calcular los elementos por terciles, cuartiles, quintiles, deciles, centiles, etc.
}
\details{
Los resultados se muestran en una matriz, donde el nombre de las filas dependerá de la elección
de p_corte, si es TRUE en los nombres de las filas se mostrará el rango de los puntos de corte de cada percentil,
si es FALSE se mostrará por el número de partes. Asimismo, si usa algunas de las funciones de la
familia `apply` el nombre de las filas no se mostrará.
}
\examples{
# El nombre de las filas en partes. Para el caso de quintiles.
set.seed(2020)
n_percentil(sample(1:100, 400,T), len=5, p_corte=F)
#    [,1]
# 1   82
# 2   82
# 3   80
# 4   77
# 5   79
# El nombre de las filas con los puntos de corte. Para el caso de quintiles.
# set.seed(2020)
n_percentil(sample(1:100, 400,T), len=5)
#           [,1]
# [1,22]     82
# (22,41]    82
# (41,58]    80
# (58,80]    77
# (80,100]   79
}
