% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{oracion}
\alias{oracion}
\title{Pasar texto a formato oración}
\usage{
oracion(x)
}
\arguments{
\item{x}{Un elemento o vector character.}
}
\description{
Esta función permite colocar texto en formato oración, como se podría hacer en un documento word.
}
\details{
Esta función se usa principalmente para la presentación de resultados.
}
\examples{
oracion(c("COMERCIO DE MANZANAS", "el ARSENAL es el mejor EQUIPO del MUNDO"))
# "Comercio de manzanas" "El arsenal es el mejor equipo del mundo"
}
