% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{limpiecito}
\alias{limpiecito}
\title{Limpieza de caracteres o strings.}
\usage{
limpiecito(x, capital = TRUE)
}
\arguments{
\item{x}{Un elemento o vector character.}

\item{capital}{Valor lógico  que toma por defecto TRUE, lo que indica que los resultados
se mostrarán en mayúsculas y si toma el valor de FALSE los resultados se mostrarán en
minúsculas.}
}
\description{
Es una función para limpiar espacios por demás al inicio, final o intermedio. Asimismo
elimina tildes y arroja los resultados en letras mayúsculas o minúsculas.
}
\details{
Esta función se usará para la limpieza de base de datos.
}
\examples{
limpiecito(c("    raul", " yo uso    R"))
# "RAUL" "YO USO R"
}
